Added validation and test detection code.
Tried to use tensorboard to check the result of the validation, but
only 1 data point is recorded. Quite wired. Also tried to use tensor
board to check the training result folder. Also very limited data.

The test result is extremely bad: so many faces everywhere. I think
the problem is the training problem.